Add GestureHandlerV1Button for RNGH v1
