+++ src/__swaps__/screens/Swap/components/controls/SwapInputAsset.tsx
@@ -7,7 +7,7 @@ import { ScreenCornerRadius } from 'react-native-screen-corner-radius';
 import { AnimatedText, Box, Column, Columns, Stack, useColorMode } from '@/design-system';
 import { useTheme } from '@/theme';
 
-import { GestureHandlerButton } from '../GestureHandlerButton';
+import { GestureHandlerV1Button } from '../GestureHandlerV1Button';
 import { SwapActionButton } from '../SwapActionButton';
 import { SwapCoinIcon } from '../SwapCoinIcon';
 import { FadeMask } from '../FadeMask';
@@ -75,7 +75,7 @@ export function SwapInputAsset() {
                 )}
               </Box>
             </Column>
-            <GestureHandlerButton
+            <GestureHandlerV1Button
               disableButtonPressWrapper
               onPressStartWorklet={() => {
                 'worklet';
@@ -98,7 +98,7 @@ export function SwapInputAsset() {
                   />
                 </Animated.View>
               </MaskedView>
-            </GestureHandlerButton>
+            </GestureHandlerV1Button>
             <Column width="content">
               <SwapActionButton
                 color={topColor}
